backend/requirements.txt
# Django and REST Framework
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=24.0

# Database
psycopg2-binary>=2.9.9

# ML and Data Processing
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
joblib>=1.3.0

# HTTP Client
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Production server
gunicorn>=21.0.0

# Development and Testing
pytest>=8.0.0
pytest-django>=4.8.0
factory-boy>=3.3.0

# Code quality
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
